backend/requirements.txt
psycopg2-binary~=2.9.6
pydantic~=1.10.9
fastapi~=0.98
uvicorn~=0.17